Optimize internal BSS table updates based on a specific BSSID
When wpa_supplicant needed to update the internal BSS table with the
latest scan results from the driver, it fetched all BSSs and processed
them all. This is unnecessary for cases where an update is needed only
for a specific BSS. Optimize this by filtering out the unnecessary
entries from the results.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>

CRs-Fixed: 3781212
Git-repo: git://w1.fi/hostap.git
Git-commit: 5b4a78b1f90977fdd013baffa0a05c47ef59f817
Change-Id: I6a115b000a234dcb6f2d462ab92d4e2701e6adec