PASN: Add set and get API for PASN data context
Modules that use libpasn for PASN authentication need the context of
PASN data. PASN data is a common context for the library and the modules
using it. Hence, initialize the context through init and deinit
functions. Also use set and get functions to update the parameters.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>

CRs-Fixed: 3781339
Git-repo: git://w1.fi/hostap.git
Git-commit: 147f83692475e1d550fc784618bc3e4be2fa8c04
Change-Id: I9c3347aa502e754008ff65d94e070c21d89c17b4