P2P2: Configure PMK/PMKID to the driver on the GO
In P2P2, the PMKSA is generated through the PASN frame exchange during
the pairing process. Once pairing and group negotiation are successful,
the P2P Client initiates a connection with the P2P GO using the PMKID
derived during the pairing process. In the case of AP SME offload to the
driver, the driver handles the Association Request and Response frames
and needs to be able to determine whether the provided PMKID is known.
Therefore, configure the PMKID to the driver on the P2P GO to allow
association with the P2P Client.

Signed-off-by: Shivani Baranwal <quic_shivbara@quicinc.com>

CRs-Fixed: 3911487
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 319c4b4e9b85f6ac40b1a33e94fc4f1504bc4736
Change-Id: Id57a9c5e39c165bfa4922e07a405e8093cb9fc11