NAN USD: Add publishChannelList option for Subscriber
Add frequency list to active NAN USD Subscriber to search for a
Publisher on multiple channels. This is the publish channel list used by
the Subscriber to periodically search for a service on these channels.
publishChannelList was already supported in the Publisher and this
commit extends that to the Subscriber.

This is needed for a P2P2 seeker that is an active subscriber looking
for an advertiser on a list of publish channels.

Signed-off-by: Shivani Baranwal <quic_shivbara@quicinc.com>

CRs-Fixed: 3911487
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 25c6598f308ea89cf9ccc90c32dc63d9c0e4efa9
Change-Id: I786be9e1c8094a4ff27fe8443729fa20bf8523c2