NAN: Handle A3 copying internally to simplify control interface
There is no need to copy the A3 value for follow-up frames through the
control interface events and commands since it can be handled internally
in the service with sufficient accuracy. More parallel operations with
multiple peers might need per-peer information, but that can be extended
in the future, if that level of complexity is really needed in practice.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>

CRs-Fixed: 3912360
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: b3bd49f3c0e3335b96c7c3f8f65e8dbe867fbf28
Change-Id: I85418e3daa91ef319e20f93a7ae310708b3eaa91