Control interface command to flush NAN publish and subscribe sessions
Add a control interface command for NAN_FLUSH to terminate all the
publish and subscribe sessions. This was previously done as a part of
the complete FLUSH operation, but it can be useful to be able to do this
specifically for NAN services without impacting other areas.

Signed-off-by: Shivani Baranwal <quic_shivbara@quicinc.com>

CRs-Fixed: 3912360
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: ae36944a9c2a57900fac58f928cf3ccadcbf57db
Change-Id: I39f673633dc0895bd4b7940c3301ead21ae13278